+++ app/controllers/devise/second_factor_webauthn_credentials_controller.rb
@@ -24,7 +24,12 @@ module Devise
     end
 
     def destroy
-      resource.second_factor_webauthn_credentials.destroy(params[:id])
+      if resource.second_factor_webauthn_credentials.destroy(params[:id])
+        set_flash_message! :notice, :security_key_deleted
+      else
+        set_flash_message! :alert, :security_key_deletion_failed, scope: :"devise.failure"
+      end
+
       redirect_to after_update_path
     end